add lazyloading and fixed bugs
