+++ src/js/catalog.js
@@ -1,12 +1,8 @@
 import { getSearchMovies, getWeeklyTrends } from './fetchmoviedata';
-import { onPagination } from './pagination';
+import { createPagination } from './pagination';
 import { createMovieCardMarkup } from './createmoviecardmarkup';
 import { warningMessageMarkup } from './createwarningmessagemurkup';
 import { refs } from './refs';
-
-import Pagination from 'tui-pagination';
-
-const container = refs.paginationEl;
 
 let page = 1;
 
@@ -24,44 +20,12 @@ export function onCatalogPage() {
 
     try {
       const videos = await getSearchMovies(query, page);
+      refs.paginationEl.classList.remove('is-hidden');
 
-      const options = {
-        // below default value of options
-        totalItems: `${videos.total_results}`,
-        itemsPerPage: `${videos.results.length}`,
-        visiblePages: 5,
-        page,
-        centerAlign: false,
-        firstItemClassName: 'tui-first-child',
-        lastItemClassName: 'tui-last-child',
-        template: {
-          page: '<a href="#" class="tui-page-btn">{{page}}</a>',
-          currentPage:
-            '<strong class="tui-page-btn tui-is-selected">{{page}}</strong>',
-          moveButton:
-            '<a href="#" class="tui-page-btn tui-{{type}}">' +
-            '<span class="tui-ico-{{type}}">{{type}}</span>' +
-            '</a>',
-          disabledMoveButton:
-            '<span class="tui-page-btn tui-is-disabled tui-{{type}}">' +
-            '<span class="tui-ico-{{type}}">{{type}}</span>' +
-            '</span>',
-          moreButton:
-            '<a href="#" class="tui-page-btn tui-{{type}}-is-ellip">' +
-            '<span class="tui-ico-ellip">...</span>' +
-            '</a>',
-        },
-      };
-      const pagination = new Pagination(container, options);
-
-      pagination.on('afterMove', event => {
-        const currentPage = event.page;
-        onPagination(query, currentPage);
-      });
-
-      console.log(videos.results);
+      createPagination(videos, query);
 
       if (videos.results.length === 0) {
+        refs.paginationEl.classList.add('is-hidden');
         renderWarningMessage();
         return;
       }
