+++ src/partials/catalog.html
@@ -19,5 +19,5 @@
   <ul class="movie-gallery__list"></ul>
   <div class="movie-gallery__message js-theme"></div>
   <!-- ==============Pagination===================== -->
-  <div id="tui-pagination-container" class="tui-pagination"></div>
+  <div id="tui-pagination-container" class="tui-pagination is-hidden"></div>
 </section>
