+++ src/sass/layout/_catalog.scss
@@ -106,6 +106,10 @@
   margin: 40px 0 0 0;
 }
 
+.is-hidden {
+  display: none;
+}
+
 .tui-page-btn,
 .tui-page-btn.tui-is-disabled,
 .tui-page-btn,
